+++ src/app/shared/shared.module.ts
@@ -5,7 +5,8 @@ import { RouterModule } from '@angular/router';
 import { RoundButton } from 'src/app/shared/components/round-button/fs-round-button.component';
 import { OfferTileComponent } from 'src/app/shared/components/offer-tile/offer-tile.component';
 import { ErrorBackTextComponent} from 'src/app/shared/components/error-back-text/error-back-text.component';
-import { LoadingSpinnerComponent } from 'src/app/shared//components/loading-spinner/loading-spinner.component';
+import { LoadingSpinnerComponent } from 'src/app/shared/components/loading-spinner/loading-spinner.component';
+import { DrawableCanvasComponent } from 'src/app/shared/components/drawable-canvas/drawable-canvas.component';
 
 import {
   FaIconLibrary,
@@ -33,7 +34,8 @@ declarations: [
   RoundButton,
   OfferTileComponent,
   ErrorBackTextComponent,
-  LoadingSpinnerComponent
+  LoadingSpinnerComponent,
+  DrawableCanvasComponent
 ],
 imports: [
   CommonModule,
@@ -48,7 +50,8 @@ exports: [
   RoundButton,
   OfferTileComponent,
   ErrorBackTextComponent,
-  LoadingSpinnerComponent
+  LoadingSpinnerComponent,
+  DrawableCanvasComponent
 ]
 })