+++ src/app/sites/ki-tools/ki-tools.module.ts
@@ -3,9 +3,11 @@ import { SharedModule } from 'src/app/shared/shared.module';
 
 import { KiToolsRoutingModule } from './ki-tools-routing.module';
 import { KiToolsComponent } from './pages/ki-tools.component';
+import { NumberCanvasComponent } from './components/number-canvas/number-canvas.component';
+import { PredictNumberDirective } from './components/number-canvas/predict-number.directive';
 
 @NgModule({
-  declarations: [KiToolsComponent],
+  declarations: [KiToolsComponent, NumberCanvasComponent, PredictNumberDirective],
   imports: [
     SharedModule,
     KiToolsRoutingModule,